refactor(): point to secondary entry points
