+++ src/app/components/content-container/content-container.module.ts
@@ -4,10 +4,11 @@ import { CommonModule } from '@angular/common';
 
 import { MatSidenavModule } from '@angular/material/sidenav';
 
-import { CovalentMediaModule, CovalentCommonModule } from '../../../platform/core';
 import { RouterModule } from '@angular/router';
 import { ContentContainerComponent } from './content-container.component';
 import { SidenavContentModule } from '../shared/sidenav-content/sidenav-content.module';
+import { CovalentMediaModule } from '../../../platform/core/media';
+import { CovalentCommonModule } from '../../../platform/core/common';
 
 @NgModule({
   declarations: [ContentContainerComponent],
