+++ src/app/components/shared/component-overview/component-overview.module.ts
@@ -7,9 +7,11 @@ import { MatListModule } from '@angular/material/list';
 
 import { ComponentOverviewComponent } from './component-overview.component';
 
-import { CovalentExpansionPanelModule, CovalentMediaModule, CovalentCommonModule } from '../../../../platform/core';
 import { RouterModule } from '@angular/router';
 import { DocumentationToolsModule } from 'app/documentation-tools';
+import { CovalentExpansionPanelModule } from '../../../../platform/core/expansion-panel';
+import { CovalentMediaModule } from '../../../../platform/core/media';
+import { CovalentCommonModule } from '../../../../platform/core/common';
 
 @NgModule({
   declarations: [ComponentOverviewComponent],
