+++ src/app/components/toolbar/toolbar.module.ts
@@ -8,8 +8,8 @@ import { MatButtonModule } from '@angular/material/button';
 import { MatListModule } from '@angular/material/list';
 import { MatIconModule } from '@angular/material/icon';
 import { MatMenuModule } from '@angular/material/menu';
-
-import { CovalentMenuModule, CovalentNotificationsModule } from '../../../platform/core';
+import { CovalentMenuModule } from '../../../platform/core/menu';
+import { CovalentNotificationsModule } from '../../../platform/core/notifications';
 
 @NgModule({
   imports: [
