+++ src/app/content/components/component-demos/layout-card-over/demos/layout-card-over-demo-basic/layout-card-over-card-over-demo.shared.ts
@@ -1,10 +1,10 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { LayoutCardOverDemoBasicComponent } from './layout-card-over-demo-basic.component';
-import { CovalentLayoutModule } from '../../../../../../../platform/core';
 import { MatIconModule } from '@angular/material/icon';
 import { MatListModule } from '@angular/material/list';
 import { MatButtonModule } from '@angular/material/button';
+import { CovalentLayoutModule } from '../../../../../../../platform/core/layout';
 
 @NgModule({
   declarations: [LayoutCardOverDemoBasicComponent],
