+++ src/app/content/components/component-demos/layout-card-over/demos/layout-card-over-demo.module.ts
@@ -2,8 +2,8 @@ import { NgModule } from '@angular/core';
 import { DemoModule } from '../../../../../components/shared/demo-tools/demo.module';
 import { LayoutCardOverDemoComponent } from './layout-card-over-demo.component';
 import { LayoutCardOverDemoRoutingModule } from './layout-card-over-demo-routing.module';
-import { CovalentLayoutModule } from '../../../../../../platform/core';
 import { LayoutCardOverDemoSharedModule } from './layout-card-over-demo-basic/layout-card-over-card-over-demo.shared';
+import { CovalentLayoutModule } from '../../../../../../platform/core/layout';
 
 @NgModule({
   declarations: [LayoutCardOverDemoComponent],
