+++ src/app/content/components/component-demos/layout-management-list/demos/layout-management-list-demo-basic/layout-management-list-demo.shared.ts
@@ -1,11 +1,11 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { LayoutManagementListDemoBasicComponent } from './layout-management-list-demo-basic.component';
-import { CovalentLayoutModule } from '../../../../../../../platform/core';
 import { MatListModule } from '@angular/material/list';
 import { MatButtonModule } from '@angular/material/button';
 import { MatIconModule } from '@angular/material/icon';
 import { MatToolbarModule } from '@angular/material/toolbar';
+import { CovalentLayoutModule } from '../../../../../../../platform/core/layout';
 
 @NgModule({
   declarations: [LayoutManagementListDemoBasicComponent],
