+++ src/app/content/components/component-demos/layout-nav-list/demos/layout-nav-list-demo-basic/layout-nav-list-demo.shared.ts
@@ -1,10 +1,10 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
-import { CovalentLayoutModule } from '../../../../../../../platform/core';
 import { LayoutNavListDemoBasicComponent } from './layout-nav-list-demo-basic.component';
 import { MatListModule } from '@angular/material/list';
 import { MatIconModule } from '@angular/material/icon';
 import { MatButtonModule } from '@angular/material/button';
+import { CovalentLayoutModule } from '../../../../../../../platform/core/layout';
 
 @NgModule({
   declarations: [LayoutNavListDemoBasicComponent],
