+++ src/app/content/components/component-demos/layout-nav-list/demos/layout-nav-list-demo.module.ts
@@ -1,9 +1,9 @@
 import { NgModule } from '@angular/core';
 import { DemoModule } from '../../../../../components/shared/demo-tools/demo.module';
-import { CovalentLayoutModule } from '../../../../../../platform/core';
 import { LayoutNavListDemoSharedModule } from './layout-nav-list-demo-basic/layout-nav-list-demo.shared';
 import { LayoutNavListDemoRoutingModule } from './layout-nav-list-demo-routing.module';
 import { LayoutNavListDemoComponent } from './layout-nav-list-demo.component';
+import { CovalentLayoutModule } from '../../../../../../platform/core/layout';
 
 @NgModule({
   declarations: [LayoutNavListDemoComponent],
