+++ src/app/content/components/component-demos/layout-nav/demos/layout-nav-demo-basic/layout-nav-demo.shared.ts
@@ -1,10 +1,10 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
-import { CovalentLayoutModule } from '../../../../../../../platform/core';
 import { LayoutNavDemoBasicComponent } from './layout-nav-demo-basic.component';
 import { MatListModule } from '@angular/material/list';
 import { MatIconModule } from '@angular/material/icon';
 import { MatButtonModule } from '@angular/material/button';
+import { CovalentLayoutModule } from '../../../../../../../platform/core/layout';
 
 @NgModule({
   declarations: [LayoutNavDemoBasicComponent],
