+++ src/app/content/components/component-demos/layout-nav/demos/layout-nav-demo.module.ts
@@ -1,9 +1,9 @@
 import { NgModule } from '@angular/core';
 import { DemoModule } from '../../../../../components/shared/demo-tools/demo.module';
-import { CovalentLayoutModule } from '../../../../../../platform/core';
 import { LayoutNavDemoComponent } from './layout-nav-demo.component';
 import { LayoutNavDemoSharedModule } from './layout-nav-demo-basic/layout-nav-demo.shared';
 import { LayoutNavDemoRoutingModule } from './layout-nav-demo-routing.module';
+import { CovalentLayoutModule } from '../../../../../../platform/core/layout';
 
 @NgModule({
   declarations: [LayoutNavDemoComponent],
