+++ src/app/content/utilities/content-details/content-details.module.ts
@@ -5,7 +5,6 @@ import { MatCardModule } from '@angular/material/card';
 import { MatIconModule } from '@angular/material/icon';
 import { MatListModule } from '@angular/material/list';
 
-import { CovalentExpansionPanelModule, CovalentMediaModule, CovalentCommonModule } from '../../../../platform/core';
 import { RouterModule } from '@angular/router';
 
 import { ContentDetailsComponent } from './content-details.component';
@@ -17,6 +16,9 @@ import { ContentExamplesComponent } from './content-examples/content-examples.co
 import { PortalModule } from '@angular/cdk/portal';
 
 import { DocumentationToolsModule } from 'app/documentation-tools';
+import { CovalentExpansionPanelModule } from '../../../../platform/core/expansion-panel';
+import { CovalentMediaModule } from '../../../../platform/core/media';
+import { CovalentCommonModule } from '../../../../platform/core/common';
 
 @NgModule({
   declarations: [ContentDetailsComponent, ContentOverviewComponent, ContentApiComponent, ContentExamplesComponent],
